requirements.txt
aniso8601==7.0.0
asgiref==3.3.1
Django==3.1.7
django-binary-database-files==1.0.14
django-cors-headers==3.7.0
django-filter==2.4.0
django-graphql-jwt==0.3.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
graphene==2.1.8
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
Markdown==3.3.4
promise==2.3
psycopg2==2.8.6
PyJWT==1.7.0
python-dateutil==2.8.1
pytz==2021.1
Rx==1.6.1
singledispatch==3.6.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
